+++ code/modules/reagents/reagent_containers/glass.dm
@@ -236,16 +236,27 @@
 
 
 /obj/item/weapon/reagent_containers/glass/barrel/attackby(var/obj/item/I, var/mob/user)
-	if (istype(I, /obj/item/weapon/reagent_containers/food/snacks/sulphur))
+	if (istype(I, /obj/item/stack/ore/sulphur))
 		reagents.add_reagent("sulfur",3)
-		qdel(I)
+		if (I.amount>1)
+			I.amount -= 1
+		else
+			qdel(I)
 		return
-	else if (istype(I, /obj/item/weapon/reagent_containers/food/snacks/saltpeter))
+		return
+	else if (istype(I, /obj/item/stack/ore/saltpeter))
 		reagents.add_reagent("potassium",3)
-		qdel(I)
+		if (I.amount>1)
+			I.amount -= 1
+		else
+			qdel(I)
 		return
-	else if (istype(I, /obj/item/weapon/reagent_containers/food/snacks/coal))
+	else if (istype(I, /obj/item/stack/ore/coal))
 		reagents.add_reagent("carbon",3)
-		qdel(I)
+		if (I.amount>1)
+			I.amount -= 1
+		else
+			qdel(I)
+		return
 		return
 	..()
